[command]Updated distro command for category sys
